removing unused code file following the refactorings.
