--- src/PitWallAcquisitionPlugin/PitWallAcquisitionPlugin/Acquisition/Leadeboards/IWebApiLeaderboardForwarderService.cs
@@ -1,7 +1,0 @@
-﻿namespace PitWallAcquisitionPlugin.Aggregations.Leadeboards
-{
-    public interface IWebApiLeaderboardForwarderService : IDataForwarderService
-    {
-
-    }
-}
