--- src/PitWallAcquisitionPlugin/PitWallAcquisitionPlugin/Acquisition/Telemetries/ITelemetryForwarderService.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace PitWallAcquisitionPlugin
-{
-    [Obsolete]
-    public interface ITelemetryForwarderService : IDataForwarderService
-    {
-    }
-}
