+++ src/PitWallAcquisitionPlugin/PitWallAcquisitionPlugin/WebApiForwarder.cs
@@ -2,17 +2,7 @@
 using FuelAssistantMobile.DataGathering.SimhubPlugin;
 using FuelAssistantMobile.DataGathering.SimhubPlugin.Logging;
 using GameReaderCommon;
-using PitWallAcquisitionPlugin.Acquisition;
-using PitWallAcquisitionPlugin.Acquisition.Repositories;
-using PitWallAcquisitionPlugin.Aggregations.Leadeboards;
-using PitWallAcquisitionPlugin.Aggregations.Telemetries;
-using PitWallAcquisitionPlugin.Aggregations.Telemetries.Aggregators;
-using PitWallAcquisitionPlugin.Aggregations.Telemetries.Repositories;
-using PitWallAcquisitionPlugin.HealthChecks;
-using PitWallAcquisitionPlugin.HealthChecks.Repositories;
 using PitWallAcquisitionPlugin.PluginManagerWrappers;
-using PitWallAcquisitionPlugin.Repositories;
-using PitWallAcquisitionPlugin.Tests.UI.Commands;
 using PitWallAcquisitionPlugin.UI.ViewModels;
 using PitWallAcquisitionPlugin.UI.Views;
 using SimHub.Plugins;
@@ -28,7 +18,6 @@ namespace PitWallAcquisitionPlugin
         private readonly IPluginRecordRepositoryFactory _pluginRecordFactory;
         private readonly IContainer _builder;
         private readonly ILogger _logger;
-        private readonly ITelemetryForwarderService _webApiForwarderService;
         private readonly IAcquisitionService _acquisitionService;
 
         public WebApiForwarder()
@@ -43,7 +32,7 @@ namespace PitWallAcquisitionPlugin
             ILogger logger,
             IPluginRecordRepositoryFactory pluginRecordFactory)
         {
-            IContainer builder = CreateBuilder(
+            IContainer builder = IocContainerInitialization.CreateBuilder(
                 logger,
                 this);
 
@@ -53,80 +42,7 @@ namespace PitWallAcquisitionPlugin
 
             _pluginRecordFactory = _builder.Resolve<IPluginRecordRepositoryFactory>();
 
-            _webApiForwarderService = _builder.Resolve<ITelemetryForwarderService>();
-
             _acquisitionService = _builder.Resolve<IAcquisitionService>();
-        }
-
-        private static IContainer CreateBuilder(
-            ILogger logger,
-            IDataPlugin plugin)
-        {
-            var containerBuilder = new ContainerBuilder();
-
-            containerBuilder.RegisterInstance(plugin).As<IPlugin>();
-
-            containerBuilder.RegisterInstance(logger);
-
-            containerBuilder.RegisterType<TelemetryLiveAggregator>()
-                .As<ITelemetryLiveAggregator>()
-                .SingleInstance();
-
-            //-------------
-
-            containerBuilder.RegisterType<RemotesRepository>()
-                .As<IRemotesRepository>()
-                .SingleInstance();
-
-            //-------------
-
-            containerBuilder.RegisterType<HealthCheckService>()
-                .As<IHealthCheckService>();
-
-            containerBuilder.RegisterType<HealthCheckRepository>()
-                .As<IHealthCheckRepository>()
-                .SingleInstance();
-
-            containerBuilder.RegisterType<PluginRecordRepositoryFactory>()
-                .As<IPluginRecordRepositoryFactory>()
-                .SingleInstance();
-
-            containerBuilder.RegisterType<WebApiTelemetryForwarderService>()
-                .As<ITelemetryForwarderService>()
-                .WithParameter("postToApiTimerHz", 1)
-                .WithParameter("autoReactivateTimer", 5000)
-                .SingleInstance();
-
-            containerBuilder.RegisterType<PluginSettingsCommandFactory>()
-                .As<IPluginSettingsCommandFactory>();
-
-            containerBuilder.RegisterType<PluginSettingsViewModel>()
-                .As<PluginSettingsViewModel>()
-                .As<IDisplayAvailability>()
-                .SingleInstance();
-
-            containerBuilder.RegisterType<SimhubPluginConfigurationRepository>()
-                .As<IPitWallConfiguration>()
-                .SingleInstance();
-
-            containerBuilder.RegisterType<MappingConfigurationRepository>()
-                .As<IMappingConfigurationRepository>();
-
-            containerBuilder.RegisterType<SettingsValidatorWrapper>()
-                .As<ISettingsValidator>();
-
-            containerBuilder.RegisterType<ForwarderServiceFactory>()
-                .As<IForwarderServiceFactory>();
-
-            containerBuilder.RegisterType<FakeLeaderboardLiveAggregator>()
-                .As<ILeaderboardLiveAggregator>();
-
-            containerBuilder.RegisterType<AcquisitionService>()
-                .As<IAcquisitionService>();
-
-            var builder = containerBuilder.Build();
-
-            return builder;
         }
 
         // ===========================================================
@@ -137,17 +53,10 @@ namespace PitWallAcquisitionPlugin
 
         public void DataUpdate(PluginManager pluginManager, ref GameData data)
         {
-            // THOUGHT: add call to service here to grab the data we want from plugin manager
             IPluginRecordRepository pluginRecordRepository
                 = _pluginRecordFactory.GetInstance(pluginManager);
 
-            /**
-             * Idea: use ioc framework to host plugin manager
-             * */
-
             _acquisitionService.HandleDataUpdate(pluginRecordRepository);
-
-            //_webApiForwarderService.HandleDataUpdate(pluginRecordRepository);
         }
 
         public void End(PluginManager pluginManager)
